1) move ExecutionOption.validate...() to HttpRequestUtils; 2) do the preprocess at flow/schedule submittion; 3) improve comments
