+++ azkaban-common/src/main/java/azkaban/executor/ExecutionOptions.java
@@ -16,29 +16,20 @@
 
 package azkaban.executor;
 
-import static azkaban.Constants.ConfigurationKeys.AZKABAN_EXECUTION_RESTART_LIMIT;
 import static azkaban.Constants.EventReporterConstants.EXECUTION_RETRIED_BY_AZKABAN;
 import static azkaban.Constants.EventReporterConstants.ORIGINAL_FLOW_EXECUTION_ID_BEFORE_RETRY;
-import static azkaban.executor.Status.RESTARTABLE_TERMINATED_STATUSES;
 
-import azkaban.Constants.FlowParameters;
 import azkaban.executor.mail.DefaultMailCreator;
 import azkaban.sla.SlaOption;
-import azkaban.utils.Props;
 import azkaban.utils.TypedMapWrapper;
 import com.google.common.collect.ImmutableMap;
 import com.google.gson.GsonBuilder;
-import com.sun.org.apache.xpath.internal.operations.Bool;
 import java.util.ArrayList;
-import java.util.Arrays;
 import java.util.Collection;
 import java.util.Collections;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
-import java.util.stream.Collectors;
-import javax.servlet.ServletException;
-import joptsimple.internal.Strings;
 
 /**
  * Execution options for submitted flows and scheduled flows
@@ -176,75 +167,6 @@ public class ExecutionOptions {
     return options;
   }
 
-  /**
-   * Validate the ExecutionOption's Flow-Parameters against the application level Properties
-   * @throws ServletException if any of the parameter is invalid
-   */
-  public void validatePreprocessFlowParameters(Props azProps)
-      throws ServletException {
-    List<String> errMsg = new ArrayList<>();
-
-    Map<String, String> flowParameters = this.getFlowParameters();
-    if (flowParameters == null || flowParameters.isEmpty()) {
-      return;
-    }
-    // if ALLOW_RESTART_ON_EXECUTION_STOPPED then ALLOW_RESTART_ON_STATUS += "EXECUTION_STOPPED,"
-    if (Boolean.parseBoolean(
-        flowParameters.getOrDefault(
-            FlowParameters.FLOW_PARAM_ALLOW_RESTART_ON_EXECUTION_STOPPED, "false"))
-    ){
-      flowParameters.put(
-          FlowParameters.FLOW_PARAM_ALLOW_RESTART_ON_STATUS,
-          Status.EXECUTION_STOPPED.name() + "," +
-              flowParameters.getOrDefault(FlowParameters.FLOW_PARAM_ALLOW_RESTART_ON_STATUS, "")
-      );
-    }
-    if (flowParameters.containsKey(FlowParameters.FLOW_PARAM_ALLOW_RESTART_ON_STATUS)) {
-      // user defined list; remove empty spaces
-      final List<String> statuses = Arrays.stream(flowParameters
-          .getOrDefault(FlowParameters.FLOW_PARAM_ALLOW_RESTART_ON_STATUS, "")
-          .split("\\s*,\\s*")).filter(s -> !s.trim().isEmpty()).collect(Collectors.toList());
-
-      for (String s : statuses) {
-        if (!RESTARTABLE_TERMINATED_STATUSES.contains(Status.valueOf(s))) {
-          errMsg.add(String.format("`%s` is not a valid restartable status, "
-              + "permitted status are %s", s, RESTARTABLE_TERMINATED_STATUSES));
-        }
-      }
-      flowParameters.put(FlowParameters.FLOW_PARAM_ALLOW_RESTART_ON_STATUS,
-          Strings.join(statuses, ","));
-    }
-    if (flowParameters.containsKey(FlowParameters.FLOW_PARAM_RESTART_COUNT)){
-      // check restart count limit
-      try {
-        final int flowRestartCountLimit = azProps.getInt(
-            AZKABAN_EXECUTION_RESTART_LIMIT, DEFAULT_FLOW_RESTART_LIMIT);
-        try {
-          final int flowRestartCount = Integer.parseInt(
-              flowParameters.getOrDefault(FlowParameters.FLOW_PARAM_RESTART_COUNT, "0"));
-          if (flowRestartCount > flowRestartCountLimit || flowRestartCount < 0){
-            errMsg.add(String.format(
-                "Invalid `" + FlowParameters.FLOW_PARAM_RESTART_COUNT + " = %d`, value should be "
-                    + "within [0, %d]", flowRestartCount, flowRestartCountLimit));
-          }
-        } catch (NumberFormatException e) {
-          errMsg.add(String.format(
-              "Invalid `" + FlowParameters.FLOW_PARAM_RESTART_COUNT + " = %s`, should be integer",
-              flowParameters.getOrDefault(FlowParameters.FLOW_PARAM_RESTART_COUNT, "0")));
-        }
-      } catch (NumberFormatException e) {
-        errMsg.add(String.format(
-            "Invalid `" + AZKABAN_EXECUTION_RESTART_LIMIT + " = %s`, should be integer",
-            azProps.get(AZKABAN_EXECUTION_RESTART_LIMIT)));
-      }
-    }
-
-    // throw exception if there's any error message
-    if (!errMsg.isEmpty()) {
-      throw new ServletException(String.format("ExecutionOptions is invalid, error reasons: %s",
-          errMsg));
-    }
-  }
 
   public void addAllFlowParameters(final Map<String, String> flowParam) {
     this.flowParameters.putAll(flowParam);
