+++ azkaban-web-server/src/main/java/azkaban/webapp/servlet/ExecutorServlet.java
@@ -1013,7 +1013,7 @@ public class ExecutorServlet extends LoginAbstractAzkabanServlet {
     final Props azProps = getApplication().getServerProps();
     try {
       options = HttpRequestUtils.parseFlowOptions(req, flowId);
-      options.validatePreprocessFlowParameters(azProps);
+      HttpRequestUtils.validatePreprocessFlowParameters(options, azProps);
     } catch (final ServletException e) {
       logger.info("parseFlowOptions failed", e);
       ret.put("error", "Error parsing flow options: " + e.getMessage());
